split off ACOM into separate file (untested)
